// src/models/IngredientModel.ts
export interface IngredientModel {
  _id: string;
  name: string;
  type: string;
  proteins: number;
  fat: number;
  carbohydrates: number;
  calories: number;
  price: number;
  image: string;
  image_mobile: string;
  image_large: string;
}

export interface IngredientsArray {
  ingredients: IngredientModel[];
}